fix: validate single relationship values
